pyproject.toml
[project]
name = "bench-memo-generator"
version = "0.1.0"
description = "Bench Memo Generator for the ND Supreme Court"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pdfminer.six>=20231228",
    "click>=8.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "diskcache>=5.6.0",
    "rich>=13.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
bench-memo = "src.cli:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*", "config*"]
